search_agent/mcp_server/requirements.txt
python-dotenv
pydantic
uvicorn
fastmcp
azure-search-documents
azure-identity